chore: relocate generated output from app/generated to /generated
