+++ lib/project.ts
@@ -1,4 +1,4 @@
-import { Prisma } from '@/app/generated/prisma/client'
+import { Prisma } from '@/generated/prisma/client'
 import { calculateMatchScore } from './matching'
 import { prisma } from './prisma'
